Not able to play 24 bit depth 96khz .wav file.
In WAV extractor read function ,As per the implementation, number
of bytes to read should be multiple of 3 for 24 bit sample stream
for the successful 24 to 16 bit conversion,otherwise total number
of bytes read cannot be consumed fully and that was causing the
glitches.
fix: increased the size of the input buffer so that it can support
8,16,and 24 bit input stream and made other changes to make sure
the input buffer will get consumed fully for all 3 formats.

Change-Id: I591a3c8306230c6663ee2cf7065ab3c16b8b8895
Author: Gurudatta Bhakte <gurudattax.bhakte@intel.com>
Signed-off-by: Gurudatta Bhakte <gurudattax.bhakte@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 28337