libmediaplayersevice: add httplive replay function for html5 to improve user experience.
It does not support httplive replay function per google code, in html5 player, the player
can not exist automatically, if the user press the play button, nothing will happen per google's code.
but if the streaming is http. not httplive ,it can replay. Add the replay func to httplive also.

Note: This patch is related to a change in platforms/frameworks/av

Change-Id: Ie1e45f7356d5a9fb7684c22d0381d4d83053cba7
Author: Yi Wang <yi.a.wang@intel.com>
Signed-off-by: Yi Wang <yi.a.wang@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 48710 28933 31510 33008 27710