MPEG4Extractor: Avoid some unexpected rounded situation when calculating sample time
Root cause:
Normally the sampletime can be divided without remainder,
a correct frame can be select as the thumbnail. But if remainder exists
in unexpected situation and the sample time will be rounded, the calculated
thumbnail sample time will smaller than the orignail and the framework will
select the previous key frame other then the correct one.

Change-Id: I3fc62f8baf4a30c6982786d2b7fac26a6f8fa4e6
Author: Weian Chen <weian.chen@intel.com>
Signed-off-by: Weian Chen <weian.chen@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 48862, 18544, 739