MPEG4Extractor: Avoid some unexpected rounded situation  when calculating sample time
Root cause: Normally the sampletime can be divided without remainder,
a correct frame can be select as the thumbnail. But if remainder exists
in unexpected situation and the sample time will be rounded, the calculated
thumbnail sample time will smaller than the orignail and the framework will
select the previous key frame other then the correct one.

Change-Id: I065443455fa563247d37525764b4c282ed553ac0
Author: Weian Chen <weian.chen@intel.com>
Signed-off-by: Weian Chen <weian.chen@intel.com>
Signed-off-by: Tiani Chen<tianmi.chen@intel.com>
Signed-off-by: Xiaowei Li<xiaowei.a.li@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 48862, 739, 18544