AwesomePlayer: Fix http streaming playback in bad network exit ANR issue
When the network is bad, it may block in request data from the web server. Especially the
AwesomePlayer in prepare status. This may cause reset ANR when awesomeplayer in prepare
status. The fix is to diconnect the mConnectionDataSource to make the blocking calls
return err when reset is called.

Change-Id: Iadedceab72df086cd6b246588a6e273a85230d93
Author: Feng Xiao<fengx.xiao@intel.com>
Signed-off-by: Feng Xiao<fengx.xiao@intel.com>
Signed-off-by: Yi Wang<yi.a.wang@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 49390, 22732, 29206