AwesomePlayer: fix http streaming playback in bad network exit ANR issue
When the network is bad, it may block in request data from the web server. Especially the
AwesomePlayer in prepare status. This may cause reset ANR when awesomeplayer in prepare
status. The fix is to diconnect the mConnectionDataSource to make the blocking calls
return err when reset is called.

Change-Id: I4926a718076304061e52088f5739b8c299edac04
Author: Feng Xiao <fengx.xiao@intel.com>
Signed-off-by: Feng Xiao <fengx.xiao@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 49390 22732 29206