To fix the green title edge problem (Native)
The problem was caused by alpha blending between some useless
green lines in the title image and the input image. In the
patch, we correct this wrong title boundary and only perform
blending between input image and title area.
There is also a very great side effect that: the computation
complexity is significantly reduced. It saves the exporting
time up to 60% for 1080P video with title effect.

Note: This patch is related to a change in platforms/frameworks/base

Change-Id: I15973f0eee972fac5e8c2b071252657d8a1b1d85
Author: Feng Wei<wei.feng@intel.com>
Signed-off-by: Feng Wei<wei.feng@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 39058