Avoid crash if sw codec is shut down while stopping
If stop() is called on a sw codec when executing the component
is requested to change state to Idle. This is done asyncronous.
When the destructor is called on the codec it is also requesting
the component to change state to Idle. If these two calls
happens soon after each other the transition may be invoked twice.
This commit will check if component is already in Idle state and
just return if it is instead of aborting.

Change-Id: I381b4aa334f17017658483e84928ef3c5bf5e142