Changed MPEG4Extractor to not look for a sync sample on audio files
When seeking in mpeg files the sample that corresponds to the
desired frame is first calculated. The closest sync sample is then
extracted from the syncsample table and that position is the place
where playback will continue. However, in an audio only MPEG4 file
the number of sync samples may be set to 0 as the concept of sync
samples does not exist in audio playback. If this value was set to
0 the file would begin playback from beginning of the file. Now it
will instead continue playing from the correct position.

Change-Id: I314a20750b3628c04a1749e091102a8d2ba38e71