StageFright: SampleTable bisection procedure correction
If searched value is between mid and right point, left
point shouldn't be set in mid + 1 because we may already
miss the searched value and that incremention contradicts
the definition of bisection procedure. Left point must
be set exactly in middle point.
In the same time, tolerance condition should be expanded
to prevent infinite loop. Actually, bisection procedure
should be stopped when left and right value distinguishes
by 1 (in this time midpoint will be calculated as equal
to left).
This correction yields in more accurate sync points
search and seek smoothing.

Change-Id: I9cf75a9a697ccb045837748e3fbba7f11756e031
Signed-off-by: Sergii Iegorov <x0155539@ti.com>