Limit read size for HLS, HTTP Live Streaming.
When streaming using HLS (HTTP Live Streaming) as opposed to
regular HTTP streaming, the sample size can become quite large
(several megabytes) corresponding to the segment size of the stream.
(With regular HTTP streaming the sample (ie read) size is 64 kBytes.)
This confuses the bandwidth estimation algorithm (in HTTPBase.cpp),
used by both HLS and HTTP streaming, and because it uses up to
100 samples, can take several minutes to adapt its estimation.
This is a severe problem during handover from 2G to 3G for example,
where the user will want a quick switch to a higher quality stream.
Limiting the size that is read, limits the sample size, and lets the
HLS bandwidth estimation adapt quickly, similar to that of
regular HTTP streaming.

Change-Id: I54ca01bbd5c6bd0438c5cbce049deae5f3d63565