Avoid appending corrupt data to 3GPP Timed text
3GPP Timed text buffers were added by the TimedText3GPPSource
with the media buffer total size instead of using correct range.

The string size is contained in the timed text buffer but
since the standard depended on the actual buffer size for additional
style information after the string the source would continue to parse
the buffer and corrupt data could be added in certain cases.

Change-Id: Ic38c4ea0182bc06b519db10e1ebad3e61cb7b2be