[AudioFlinger]rework of remove_effect calls
Bug issue is (probably, because unable to reproduce) caused by a
desynchronization of threads: a thread indirectly frees memory of
an "input stream" object then another thread continues to use
"intput stream" object. Because these threads do not share Mutex::Autolock
object, the solution to issue is rework of "input stream" usage. I.e.
currently, a thread calls a first time a function of "input stream" and
later, the same threads recalls a second time the same function of "input
stream". This second call is useless and can be avoided.

So the purpose of this patch is
 - addition of a boolean attribute to "thread" which indicates if
"input stream" function has been called or not
 - pooling of calls to "input stream" function following boolean

Change-Id: I9a80246d2d4234b332dc9f0f1cc0fe4bbd763482
Author: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Author-tracking-BZ: 61296