Avoid mediaserver crash when failing to set input port size
The mediaserver will crash when passing an invalid input port size
to the OMX codec plugins. For example, if the sample size in stsz
table is negtive, this will make kKeyMaxInputSize value very large.
The value will be passed to OMX plugins to set the input port size.
OMX plugins will detect this invalid value and report errors. But
the return error is not properly handled in setMinBufferSize and
the CHECK() will crash the mediaserver directly.

This patch will handle this situation and allows OMXCodec to exit
elegantly.

Change-Id: Ie5bc2ba15f158ac2a4fc448017845b956f137aff
Author: Wei Feng <wei.feng@intel.com>
Signed-off-by: Wei Feng <wei.feng@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 68835