Failed to play multichannel audio repeatly with audio direct output
This issue caused by the gapless playback feature, which is no fit for
multichannel audio.When the track use the direct output, it can not be used for recycle.
After finishing playback , the direct output must be released, and
in beginning of playback it will be recreated. In cycle playback, if the next
track also need to create direct output, but previous one is not released, it will cause issue.
Here we checking the output flag, if it is direct output, release the track first.

Change-Id: I4c2c6352003c979cb626e392688e5b932f54c52f
Signed-off-by: b02247 <b02247@freescale.com>
Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>