Enable NuPlayer getMetadata in prepared state
According to the MediaPlayer api documentation the metadata
about the stream and the duration should be available when
the player is in the prepared state. This commit will initialize
the connection to the server when setDataSource is called and
block the prepare call until the connection is established and
metadata and duration information is available. It will also
implement the getMetadata function to return can pause and can
seek for the RTSP and HLS sources.

Change-Id: I8716e62025346ed989e0a876731e7845d2585b9d