Progressbar fix when seeking in NuPlayer
When a seek is performed in NuPlayer the progressbar is updated to
the new position for a short while, but then jumps back to the
previous position when old buffer units are being rendered.
Then it jumps to the new position again when the new frames are
rendered.

This fix makes sure that the new seek position is displayed until
the player starts rendering from that position, i.e. the jump back
to the previous position is removed.

Change-Id: I09cda5e67795ffa039345b3032e110ee1cee0488