Fix 16-byte alignment issue when HDCP is used over WFD
According to HDCP spec. rev 2.1, "3.6.2 Transport Stream",
the 16 byte encryption block boundary must be aligned with the
start of a PES payload. This patch adds appropriate size of
paddings in the Adaptation Field to meet such requirement.
It should improve interoperability with Wi-Fi Display sinks
that strictly follow HDCP spec.

Change-Id: I91b7bc28cf1884bec7e5c8c2b504814881d9932e
Signed-off-by: Sho Amano <sho.amano@gmail.com>