Don't change value when getting AudioParameters fails
When calling getInt or getFloat of AudioParameters and the call fails
don't change the input variable. Previously it was set to 0 no matter
if the call was successful or failed which was not very intuitive.

getInt("Key", &value) would for example change value to 0 even if
"Key" was not present as an AudioParameter.

Change-Id: Ie7f12afb2c6188d94558f95e78e6612a81db9fd7