avcenc: Include more parameters in getParam(OMX_IndexParamPortDefinition)
The SimpleSoftOMXComponent base class only updates the
nBufferSize and nBufferCountActual parameters in internalSetParameter,
so only these parameters have sensible values when queried.

Make sure the eColorFormat parameter is set to the actual
value that is set up, and return OMX_COLOR_FormatAndroidOpaque
instead of YUV420SemiPlanar if that has been set up. (Also
make sure to keep the internal color format as YUV420SemiPlanar
after a setParam(OMX_IndexParamPortDefinition) call that
sets the color format to OMX_COLOR_FormatAndroidOpaque.)

This fixes failures in OMXNodeInstance::createInputSurface
with this encoder - it seems to be able to handle surface input
just fine once the caller trusts it enough to proceed.

Also return the actual frame width/height on the output port.
This makes sure the correct width gets included in the
output format MediaFormat from the MediaCodec encoder.

Change-Id: I0667b3b4c6bb90331ad0ae7d51388e1bca3d1bbd