avcenc: Include more parameters in getParam(OMX_IndexParamPortDefinition)
The SimpleSoftOMXComponent base class only updates the
nBufferSize and nBufferCountActual parameters in internalSetParameter,
so only these parameters have sensible values when queried.
Additionally we return the actual frame width/height on both ports,
and the right color format. This makes sure that the output MediaFormat
from the MediaCodec encoder contains the right size parameters (which
previously always stayed as 176x144).

(Also make sure to keep the internal color format as YUV420SemiPlanar
after a setParam(OMX_IndexParamPortDefinition) call that sets the
color format to OMX_COLOR_FormatAndroidOpaque.)

This fixes failures in OMXNodeInstance::createInputSurface with this
encoder, although this is not enough to make surface input work
properly (tested on a Nexus 4).

Change-Id: I0667b3b4c6bb90331ad0ae7d51388e1bca3d1bbd