Remove gap when switching to and from background playback
If an application wants to play the sound of a video in the
background, it can signal the transitions to the framework by calling
setDisplay(). In Awesomeplayer when this happens both audio and video
will be paused during the shutdown and reinitialization of the video
decoder. When the new video decoder is up and running, the player will
seek to the last shown video position. This will cause a small gap in
the audio playback. This is fixed by only pausing and seeking the
video and letting the audio continue to play.

Change-Id: I8bb75c634c06502a97bf9f199d75b5c129802a1d