Enable background playback for protected content
If an application wants to play the sound of the video in the
background, it can signal the transitions to the framework by
calling setDisplay(). Stagefright will close the video codec and
recreate it again with the new display. If background play is
activated, the display would be null. When doing this for
protected content, the initialisation of the codec would fail
because protected content needs a native window to get protected
buffers. This is solved by creating a dummy video decoder, that
will continue to send buffers with correct timestamps, when there
is no display to show the content.

Change-Id: Ifd3e51b07b6e50987c34c8bf79e2646f223cd28b