AudioPlayer: Fixed progress bar non-stop issue when close A2DP during Music playback
Issue Description:
During Music playback with BT A2DP, after close BT, sometimes the Music is
paused but the Music progress bar is still moving forward and updating.

Root Cause:
This is a side effect of patch (commit a99a5bca). That patch's purpose is
to only allow update the progress bar when buffer is being filled. However,
as Audio Track data filling and control operations are in different threads,
so in some corner case the fillBuffer() is invoked after pause() invoked
and that causes the progress bar updates continuously due to mPinnedTimeUs
is set to -1LL.

Solution:
DO NOT update the progress bar any more when Music is paused or reached EOS.

Change-Id: I66fe36e607137897babd21c4ef853141101f709d
Signed-off-by: Chao Deng <cdeng@marvell.com>